requirements.txt
absl-py==0.15.0
argon2-cffi==21.1.0
astor==0.8.1
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
backports.lzma==0.0.14
bleach==4.1.0
cached-property==1.5.2
cffi==1.15.0
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
defusedxml==0.7.1
entrypoints==0.3
gast==0.5.2
grpcio==1.41.1
h5py==2.10.0
imageio==2.10.1
importlib-metadata==4.8.1
ipykernel==5.5.6
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
Keras==2.0.8
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
mock==4.0.3
nbclient==0.5.4
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
notebook==6.4.5
numpy==1.19.5
opencv-python==4.5.4.58
packaging==21.2
pandocfilters==1.5.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prometheus-client==0.12.0
prompt-toolkit==3.0.21
protobuf==3.19.1
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
PyWavelets==1.1.1
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.2
scikit-image==0.17.2
scipy==1.2.2
Send2Trash==1.8.0
six==1.16.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
tifffile==2020.9.3
tornado==6.1
traitlets==4.3.3
typing-extensions==3.10.0.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.5.2
zipp==3.6.0
tensorflow-gpu==1.13.1